% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{ARC.IDENT}
\alias{ARC.IDENT}
\title{Spatial Identity Operation (ArcGIS Identity Equivalent)}
\usage{
ARC.IDENT(layer_a, layer_b)
}
\arguments{
\item{layer_a}{An sf object representing the input layer to be split}

\item{layer_b}{An sf object representing the overlay layer}
}
\value{
An sf object containing:
\itemize{
\item All intersecting areas between layer_a and layer_b (with attributes from both)
\item All non-intersecting areas from layer_a (with only layer_a attributes)
}
}
\description{
Performs a spatial identity operation similar to ArcGIS's Identity tool.
Computes the intersection of two layers and preserves the portions of
layer_a that do not overlap with layer_b. The output contains all areas
from layer_a, with attributes from layer_b where they overlap.
}
\details{
This is commonly used for overlay analysis where you want to split polygons
in one layer by boundaries in another layer while keeping all original areas.
}
\examples{
\dontrun{
# Split wildlife range by land ownership to get area by owner
caribou_range \%>\%
  ARC.IDENT(land_ownership)
}

}
