% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{CALC_HA}
\alias{CALC_HA}
\title{Calculate Area in Hectares}
\usage{
CALC_HA(.data)
}
\arguments{
\item{.data}{An sf object with polygon geometries}
}
\value{
The input sf object with an added "Hectares" column containing
the area of each feature in hectares
}
\description{
Adds a column containing the area of each feature in hectares. Works with
any sf object containing polygon or multipolygon geometries. Uses the
coordinate reference system of the input data, so ensure data is in an
appropriate projected CRS (not lat/long) for accurate area calculations.
}
\examples{
\dontrun{
# Add hectare areas to forest cutblocks
cutblocks_sf \%>\%
  st_transform(3005) \%>\%  # BC Albers for accurate area
  CALC_HA()
}

}
