% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{CHECK_GEOMETRY}
\alias{CHECK_GEOMETRY}
\title{Check Geometry Types in an SF Object}
\usage{
CHECK_GEOMETRY(x)
}
\arguments{
\item{x}{An sf object to check}
}
\value{
A tibble showing the count of each geometry type (POINT, LINESTRING,
POLYGON, MULTIPOLYGON, etc.)
}
\description{
Provides a quick summary count of geometry types present in an sf object.
Useful for diagnosing mixed geometry issues or verifying expected geometry types.
}
\examples{
\dontrun{
# Check what geometry types are in your spatial layer
wildlife_habitat \%>\% CHECK_GEOMETRY()
# Output might show: POLYGON (150), MULTIPOLYGON (3)
}

}
