% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.common.functions-package.R
\docType{package}
\name{my.common.functions-package}
\alias{my.common.functions-package}
\alias{my.common.functions}
\title{my.common.functions: Spatial Data Utilities and Wildlife Analysis Tools}
\description{
A collection of utility functions for exporting spatial data to themed KML/KMZ
files and creating georeferenced PDF maps with basemaps. Supports point, line,
and polygon geometries with customizable styling.
}
\section{Main Functions}{

\itemize{
\item \code{\link{EXPORT_SF_TO_THEMED_KML}}: Export sf objects to themed KML/KMZ files
\item \code{\link{CREATE_BASEMAP_PLOT}}: Create ggplot maps with integrated basemaps
\item \code{\link{EXPORT_GEOREFERENCED_PDF}}: Convert ggplot maps to georeferenced PDFs
}
}

\section{KML/KMZ Export}{

The \code{EXPORT_SF_TO_THEMED_KML} function creates Google Earth compatible files
with professional styling including categorical or continuous color schemes,
customizable labels, and support for points, lines, and polygons.
}

\section{Georeferenced Maps}{

The \code{CREATE_BASEMAP_PLOT} and \code{EXPORT_GEOREFERENCED_PDF} functions work
together to create publication-quality georeferenced PDF maps with basemaps from
ESRI, OpenStreetMap, and other providers.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bevann/my_common_functions}
  \item Report bugs at \url{https://github.com/Bevann/my_common_functions/issues}
}

}
\author{
Bevan Ernst
}
\keyword{GIS}
\keyword{KML}
\keyword{maps}
\keyword{spatial}
